requirements.txt
azure-functions==1.15.0
azure-cosmos==4.3.0
